Move box-backend into another package
This makes the qabelbox application compatible with the current
qabel-core/master and does not use the qabel.core package for code anymore.
