--- qabelbox/src/main/java/de/qabel/core/exceptions/QblStorageException.java
@@ -1,11 +1,0 @@
-package de.qabel.core.exceptions;
-
-public class QblStorageException extends QblException{
-	public QblStorageException(Throwable e) {
-		super(e);
-	}
-
-	public QblStorageException(String s) {
-		super(s);
-	}
-}
