+++ qabelbox/src/main/java/de/qabel/qabelbox/activities/MainActivity.java
@@ -31,14 +31,13 @@ import java.io.InputStream;
 import java.io.OutputStream;
 import java.net.URLConnection;
 import java.util.ArrayList;
-import java.util.Collections;
 
-import de.qabel.core.exceptions.QblStorageException;
-import de.qabel.core.storage.BoxFile;
-import de.qabel.core.storage.BoxFolder;
-import de.qabel.core.storage.BoxNavigation;
-import de.qabel.core.storage.BoxObject;
-import de.qabel.core.storage.BoxVolume;
+import de.qabel.qabelbox.exceptions.QblStorageException;
+import de.qabel.qabelbox.storage.BoxFile;
+import de.qabel.qabelbox.storage.BoxFolder;
+import de.qabel.qabelbox.storage.BoxNavigation;
+import de.qabel.qabelbox.storage.BoxObject;
+import de.qabel.qabelbox.storage.BoxVolume;
 import de.qabel.qabelbox.QabelBoxApplication;
 import de.qabel.qabelbox.R;
 import de.qabel.qabelbox.adapter.FilesAdapter;
