+++ qabelbox/src/main/java/de/qabel/qabelbox/exceptions/QblStorageException.java
@@ -1,0 +1,13 @@
+package de.qabel.qabelbox.exceptions;
+
+import de.qabel.core.exceptions.QblException;
+
+public class QblStorageException extends QblException {
+	public QblStorageException(Throwable e) {
+		super(e.getMessage());
+	}
+
+	public QblStorageException(String s) {
+		super(s);
+	}
+}
