+++ qabelbox/src/main/java/de/qabel/qabelbox/exceptions/QblStorageNotFound.java
@@ -1,4 +1,4 @@
-package de.qabel.core.exceptions;
+package de.qabel.qabelbox.exceptions;
 
 public class QblStorageNotFound extends QblStorageException {
