+++ qabelbox/src/main/java/de/qabel/qabelbox/fragments/FilesFragment.java
@@ -13,7 +13,7 @@ import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
 
-import de.qabel.core.storage.BoxNavigation;
+import de.qabel.qabelbox.storage.BoxNavigation;
 import de.qabel.qabelbox.R;
 import de.qabel.qabelbox.adapter.FilesAdapter;
