+++ qabelbox/src/main/java/de/qabel/qabelbox/fragments/SelectUploadFolderFragment.java
@@ -10,7 +10,7 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
 
-import de.qabel.core.storage.BoxNavigation;
+import de.qabel.qabelbox.storage.BoxNavigation;
 import de.qabel.qabelbox.R;
 import de.qabel.qabelbox.adapter.FilesAdapter;
