+++ qabelbox/src/main/java/de/qabel/qabelbox/storage/BoxFolder.java
@@ -1,4 +1,4 @@
-package de.qabel.core.storage;
+package de.qabel.qabelbox.storage;
 
 import java.util.Arrays;
