+++ qabelbox/src/main/java/de/qabel/qabelbox/storage/BoxObject.java
@@ -1,4 +1,4 @@
-package de.qabel.core.storage;
+package de.qabel.qabelbox.storage;
 
 public class BoxObject implements Comparable<BoxObject>  {
 	public String name;
